requirements.txt
pytest
Appium-Python-Client
python-dotenv
selenium
pytz
requests
opencv-python